Fix typecast problems.

+++ src/Packet.c
@@ -36,7 +36,7 @@ CD_PacketFromBuffer (CDBuffer* input)
     }
 
     self->chain = CDRequest;
-    self->type  = (uint32_t) CD_BufferRemoveByte(input);
+    self->type  = (uint32_t) (uint8_t) CD_BufferRemoveByte(input);
     self->data  = CD_GetPacketDataFromBuffer(self, input);
 
     if (!self->data) {
